% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_model.R
\name{cfa_model}
\alias{cfa_model}
\title{Confirmatory Factor Analysis (CFA) Model Class}
\description{
Internal R6 class to extract and organize CFA model indicators.
}
\details{
Parses lavaan model syntax specific to CFA structure and separates
fixed vs. free indicators per factor.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lavaan_syntax}}{Lavaan model syntax.}

\item{\code{fact_names}}{A character vector of factor names}

\item{\code{free_ind}}{A named list of free indicators per factor}

\item{\code{fixed_ind}}{A named list of fixed indicators per factor}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cfa_model-new}{\code{cfa_model$new()}}
\item \href{#method-cfa_model-clone}{\code{cfa_model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cfa_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-cfa_model-new}{}}}
\subsection{Method \code{new()}}{
Initialize the cfa_model object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cfa_model$new(lavaan_model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lavaan_model}}{A string in lavaan model syntax.
Sorts indicators of each factor in fixed (first) and free}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cfa_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-cfa_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cfa_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
