% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm_correction.R
\name{create_correction_element}
\alias{create_correction_element}
\title{Factor Score Correction Creation}
\usage{
create_correction_element(var_1, vals_1, var_2, vals_2, factor_names, sigma_ee)
}
\arguments{
\item{var_1}{The name of the first covariate
(either main effect or part of an interaction).}

\item{vals_1}{The predictor matrix of the first term.}

\item{var_2}{The name of the second covariate
(either main effect or part of an interaction).}

\item{vals_2}{The predictor matrix of the second term.}

\item{factor_names}{A character vector of factor (latent) variable names.}

\item{sigma_ee}{A matrix of factor score covariances,
indexed by the factor variable names.}
}
\value{
List containing the value and the correction of the element.
}
\description{
Internal function for calculating corrected elements in latent
regression models based on main effects and interaction terms.
}
\keyword{internal}
