% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\name{data_loader}
\alias{data_loader}
\title{Data Loader Class}
\description{
Internal R6 class for loading and bootstrapping input data.
}
\details{
Validates and stores input data, manages variable names,
and generates bootstrap samples.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{original_data}}{A matrix or data.frame of the input data.}

\item{\code{name_config}}{An instance of the \code{var_name_config} class.}

\item{\code{boot_samples}}{A list of bootstrap samples
generated from \code{original_data}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-data_loader-new}{\code{data_loader$new()}}
\item \href{#method-data_loader-bootstrap}{\code{data_loader$bootstrap()}}
\item \href{#method-data_loader-clone}{\code{data_loader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_loader-new"></a>}}
\if{latex}{\out{\hypertarget{method-data_loader-new}{}}}
\subsection{Method \code{new()}}{
Initialize the data_loader object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_loader$new(input_data, treatment_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_data}}{A matrix or data.frame with named columns
containing the input data.}

\item{\code{treatment_names}}{A character vector of treatment variable names.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_loader-bootstrap"></a>}}
\if{latex}{\out{\hypertarget{method-data_loader-bootstrap}{}}}
\subsection{Method \code{bootstrap()}}{
Generate bootstrap samples from the original data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_loader$bootstrap(n_bootstrap)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_bootstrap}}{Integer, number of bootstrap samples to generate.
Validates the format and column names of the input data.
Validates that treatment variables have binary format.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_loader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-data_loader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_loader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
