% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm_correction.R
\name{extract_interaction_parts}
\alias{extract_interaction_parts}
\title{Extract Interaction Parts}
\usage{
extract_interaction_parts(inter, factor_names)
}
\arguments{
\item{inter}{The interaction term to extract parts from.}

\item{factor_names}{A character vector of factor (latent) variable names.}
}
\value{
A list with two components: \code{latent} (latent variables) and
\code{observed} (observed variables).
}
\description{
Extracts the latent and observed components from a given interaction term.
An interaction term can only contain one latent variable.
}
\keyword{internal}
