% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_score_estimator.R
\name{factor_score_estimator}
\alias{factor_score_estimator}
\title{Factor Score Estimator Class}
\description{
Internal R6 class for computing factor scores from lavaan cfa output
according to the method of moments technique of Wall and Amemiya (2003).
}
\details{
Takes a \code{cfa_model} object and a data sample as input, uses the lavaan
\code{cfa} function with MLR estimation to fit the CFA model to the data
and computes the factor scores as well as the the error covariance matrix
based on extracted intercepts, loadings, and residual variances with
a method of  moments approach.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{names}}{Character vector of latent factor names.}

\item{\code{vals}}{Matrix of computed factor scores for each observation.}

\item{\code{sigma}}{Covariance matrix of factor score estimation errors.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-factor_score_estimator-new}{\code{factor_score_estimator$new()}}
\item \href{#method-factor_score_estimator-clone}{\code{factor_score_estimator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-factor_score_estimator-new"></a>}}
\if{latex}{\out{\hypertarget{method-factor_score_estimator-new}{}}}
\subsection{Method \code{new()}}{
Initializes a \code{factor_score_estimator} object and computes factor scores
from a CFA model and data sample.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{factor_score_estimator$new(cfa_model, data_sample)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cfa_model}}{An instance of \code{cfa_model} defining the
factor structure.}

\item{\code{data_sample}}{A data frame containing observed indicator values.
Core logic for estimating factor scores from a CFA model and data sample.
Fit cfa model, check if it converged and return coefficients
Extract factor loading matrix from lavaan fit
Extract residual covariance matrix from lavaan fit}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-factor_score_estimator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-factor_score_estimator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{factor_score_estimator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
