% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm_correction.R
\name{handle_double_interaction}
\alias{handle_double_interaction}
\title{Handle Double Interaction}
\usage{
handle_double_interaction(var_1, vals_1, var_2, vals_2, factor_names, sigma_ee)
}
\arguments{
\item{var_1}{The name of the first interaction term.}

\item{vals_1}{The predictor matrix of the first term.}

\item{var_2}{The name of the second interaction term.}

\item{vals_2}{The predictor matrix of the second term.}

\item{factor_names}{A character vector of factor (latent) variable names.}

\item{sigma_ee}{A matrix of factor score covariances.}
}
\value{
List containing the value and the correction of the
double interaction.
}
\description{
Calculates the corrected vector for a pair of interaction terms.
Returns the mean of their product, which is corrected with the
corresponding value from the sigma_ee matrix if both variables
contain a name found in factor_names.
}
\keyword{internal}
