% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm_correction.R
\name{handle_interaction}
\alias{handle_interaction}
\title{Handle Interaction}
\usage{
handle_interaction(
  inter,
  vals_inter,
  single,
  vals_single,
  factor_names,
  sigma_ee
)
}
\arguments{
\item{inter}{The interaction term to handle.}

\item{vals_inter}{The predictor matrix of the interaction term.}

\item{single}{The single variable (either latent or observed).}

\item{vals_single}{The predictor matrix of the single term.}

\item{factor_names}{A character vector of factor (latent) variable names.}

\item{sigma_ee}{A matrix of factor score covariances.}
}
\value{
List containing the value and the correction of the
interaction.
}
\description{
Calculates the corrected vector for a main effect and an
interaction term. Returns the mean of their product, which is corrected
with the corresponding value from the sigma_ee matrix if both variables
contain a name found in factor_names.
}
\keyword{internal}
