% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_est.R
\name{med_est}
\alias{med_est}
\title{Mediation Estimation for Latent Variables}
\usage{
med_est(
  lavaan_txt,
  data,
  treat_names,
  nboot = 100,
  modified_2smm = TRUE,
  tau = 5,
  add_var = 1e-04
)
}
\arguments{
\item{lavaan_txt}{lavaan model string}

\item{data}{input data with column names}

\item{treat_names}{character vector of treatment names}

\item{nboot}{number of bootstrap replicates; defaults to 100.}

\item{modified_2smm}{Boolean. Turns on modified 2SMM estimation.
Default is TRUE.}

\item{tau}{Numeric. Modification factor. Default is 0.}

\item{add_var}{Numeric. Small variance value added to improve
stability in g-estimation. Default is 0.}
}
\description{
Performs mediation estimation using standard factor
regression and g-estimation for unbiased estimates in the
presence of counfounding with error estimates via bootstrap.
}
