% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_corrector.R
\name{reg_corrector}
\alias{reg_corrector}
\title{R6 Class for Regression Correction based on G-Estimation}
\description{
Internal R6 class that implements correction of regression weights and
adjustment terms using factor scores, particularly in the context of
mediation or structural modeling.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{corrected_reg_estimator}}{An updated regression estimator object
with applied corrections.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-reg_corrector-new}{\code{reg_corrector$new()}}
\item \href{#method-reg_corrector-g_estimation}{\code{reg_corrector$g_estimation()}}
\item \href{#method-reg_corrector-clone}{\code{reg_corrector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reg_corrector-new"></a>}}
\if{latex}{\out{\hypertarget{method-reg_corrector-new}{}}}
\subsection{Method \code{new()}}{
Initializes the \code{reg_corrector} by applying factor score corrections
to the specified regression estimator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reg_corrector$new(
  reg_est,
  reg_model,
  sample,
  var_names,
  factor_scores,
  modified_2smm,
  tau
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reg_est}}{An instance of \code{reg_param_estimator}, containing
methods for prediction and regression.}

\item{\code{reg_model}}{An instance of \code{reg_model}, containing the
mediator predictors.}

\item{\code{sample}}{A data frame representing the sample data.}

\item{\code{var_names}}{An instance of \code{var_names_config},
containing the variable names.}

\item{\code{factor_scores}}{An instance of \code{factor_score_estimator},
containing the factor scores and covariances.}

\item{\code{modified_2smm}}{Boolean. Turns on modified 2SMM estimation.}

\item{\code{tau}}{Numeric. Modification factor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reg_corrector-g_estimation"></a>}}
\if{latex}{\out{\hypertarget{method-reg_corrector-g_estimation}{}}}
\subsection{Method \code{g_estimation()}}{
Performs the final regression estimation using the corrected weights and
adjustment terms.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reg_corrector$g_estimation(modified_2smm, tau, add_var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modified_2smm}}{Boolean. Turns on modified 2SMM estimation.}

\item{\code{tau}}{Numeric. Modification factor.}

\item{\code{add_var}}{Numeric. Small value added to the diagonal of the matrix
to improve stability (Ridge regularization).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Regression coefficients obtained with g-estimation.
Correct the regression through g-estimation weights.
Applies centering to treatment variables.
Computes the corrected entries for treatment predictors.
Computes the corrected entries for mediator predictors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reg_corrector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-reg_corrector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reg_corrector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
