% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_model.R
\name{reg_model}
\alias{reg_model}
\title{Regression Model Class}
\description{
Internal R6 class to parse and organize regression model predictors.
}
\details{
Processes lavaan regression statements to identify mediators and outcomes.
Adds intercepts where not explicitly omitted.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{text}}{A character vector of regression model lines.}

\item{\code{med_preds}}{A named list of mediator variables and their predictors.}

\item{\code{out_preds}}{A named list of outcome variables and their predictors.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-reg_model-new}{\code{reg_model$new()}}
\item \href{#method-reg_model-clone}{\code{reg_model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reg_model-new"></a>}}
\if{latex}{\out{\hypertarget{method-reg_model-new}{}}}
\subsection{Method \code{new()}}{
Initialize the regression model object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reg_model$new(reg_model_lines, intercept_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reg_model_lines}}{A character vector of regression syntax lines.}

\item{\code{intercept_name}}{Name used to represent the intercept term.
Extracts predictors for each regression equation and identifies
mediators and outcomes based on appearance in left- and right-hand sides.
Determines whether a regression line specifies a zero intercept.
Checks if a line is a valid regression line.
Parses a regression line into its outcome and covariates,
handling the intercept and interaction terms.
Updates the regression info list by merging new covariates.
Classifies variables into mediators and outcomes based on LHS/RHS usage.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reg_model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-reg_model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reg_model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
