% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_param_estimator.R
\name{reg_param_estimator}
\alias{reg_param_estimator}
\title{Regression Parameter Estimator Class}
\description{
Internal R6 class for performing regression with factor score corrections.
The class stores target and predictor variables, calculates factor
corrections, and performs the regression to estimate model parameters.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{target}}{Numeric vector of the target variable.}

\item{\code{var_target_error}}{Numeric value containing the error variance
the (latent) target variable}

\item{\code{predictors}}{Matrix of predictor variables.}

\item{\code{weights}}{Optional matrix of weights for regression.}

\item{\code{mat_m1}}{Regression matrix without corrections.}

\item{\code{vec_m1}}{Regression vector without corrections.}

\item{\code{mat_m2}}{Regression matrix factor score corrections.}

\item{\code{vec_m2}}{Regression vector factor score corrections.}

\item{\code{parameters}}{Estimated regression parameters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-reg_param_estimator-new}{\code{reg_param_estimator$new()}}
\item \href{#method-reg_param_estimator-regress}{\code{reg_param_estimator$regress()}}
\item \href{#method-reg_param_estimator-clone}{\code{reg_param_estimator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reg_param_estimator-new"></a>}}
\if{latex}{\out{\hypertarget{method-reg_param_estimator-new}{}}}
\subsection{Method \code{new()}}{
Initialize the regression parameter estimator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reg_param_estimator$new(
  sample,
  var_names,
  target_name,
  factor_scores,
  predictor_names
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample}}{Data frame containing the observed data.}

\item{\code{var_names}}{An instance of \code{var_names_config},
containing the variable names.}

\item{\code{target_name}}{String specifying the name of the target variable.}

\item{\code{factor_scores}}{An instance of \code{factor_score_estimator},
containing the factor scores and covariances.}

\item{\code{predictor_names}}{Character vector specifying the names of the
predictor variables.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reg_param_estimator-regress"></a>}}
\if{latex}{\out{\hypertarget{method-reg_param_estimator-regress}{}}}
\subsection{Method \code{regress()}}{
Perform regression on the target variable using the predictors
and weights
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reg_param_estimator$regress(modified_2smm, tau, add_var = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{modified_2smm}}{Boolean. Turns on modified 2SMM estimation.}

\item{\code{tau}}{Numeric. Modification factor.}

\item{\code{add_var}}{Numeric. Small value added to the diagonal of the matrix
to improve stability (Ridge regularization). Default is 0.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Regression coefficients.
Perform the regression calculation to estimate parameters.
Solves the system of equations using either the regular
\code{solve} function or the generalized inverse (\code{ginv}) from
the \code{MASS} package if the matrix is singular.
Center values if necessary (if variable is not
a treatment or intercept and not binary).
Prepare the matrix with centered predictors,
including interaction terms.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reg_param_estimator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-reg_param_estimator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{reg_param_estimator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
