% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_est.R
\name{run_bootstrap_reps}
\alias{run_bootstrap_reps}
\title{Run Bootstrap Replications for Mediation Estimation}
\usage{
run_bootstrap_reps(
  nboot,
  bootstrapped_samples,
  models,
  data_info,
  modified_2smm,
  tau,
  add_var
)
}
\arguments{
\item{nboot}{Integer; number of bootstrap replicates.}

\item{bootstrapped_samples}{List of bootstrapped data frames.}

\item{models}{Object containing CFA and regression models.}

\item{data_info}{Object containing data name configuration.}

\item{modified_2smm}{Boolean. Turns on modified 2SMM estimation.}

\item{tau}{Numeric. Modification factor.}

\item{add_var}{Numeric. Small variance value added to improve
stability in g-estimation.}
}
\value{
A list containing the results
}
\description{
Performs bootstrap replications of mediation estimation using
factor score estimation, standard regression, and g-estimation.
}
\keyword{internal}
